Centralize 6 function prototypes into shared headers
- Entity_ActivateHitSlot → include/entity.h
  Canonical: void*(ent), void*(mask), u8(kind), u8(tile); binary entry
  masks r2/r3 with lsls/lsrs #24 confirming u8 scalars. Def changed
  from struct EntryB8A8*/ull* to void* to match callers' pointer-passing
  convention; 8 C callers de-externalised.

- Entity_SpawnFromRecord → include/entity.h
  Canonical: s8(id); binary entry does lsls/asrs #24 (signed). Def
  fixed from u8 to s8; s32-typed caller (sub_08009ba0) de-externalised.

- EntityScript_Advance → include/entity.h
  Canonical: s8(a), s8(b), s32(c) per def. Two u32/unsigned-int callers
  de-externalised.

- Scene12_UpdateTileInteraction → include/game.h
  Canonical: void*(buf), void*(r4_obj); binary loads r0/r1 into r6/r7
  as address registers, no entry mask. Def changed from u32 to void*.
  u32-typed caller fixed with (void*) casts.

- Player_HandleTileTransitions → include/game.h
  Same pointer pattern as Scene12. Def changed from u32 to void*.
  u32-typed caller fixed with (void*) casts.

- sub_08012BC4 → include/gfx.h
  Canonical: u32 scalar params (agbcc generates the entry lsls/lsrs
  masks only for u32 params; declaring narrower would remove them and
  break the match), void* for srcTable (accommodates both void*-tiles
  and const u16** callers without casts). 14 callers de-externalised or
  proto aligned; NON_MATCHING scaffold in sub_080132e8 cleaned up.

make check passes; all 6 definitions + all checked callers byte-match.

+++ src/game/sub_0800411c.c
@@ -21,7 +21,7 @@ extern u32 Scene_InitScan(u8 *localState);
 extern u16 Input_Poll(void);
 extern void Scene_LoadBg(void);
 extern void Entity_RunScript(u32 a, u32 b);
-extern void EntityScript_Advance(u32 a, u32 b, u32 c);
+extern void EntityScript_Advance(s8 a, s8 b, s32 c);
 extern void Game_RunEntityFrame(void);
 extern void WaitVblank(void);
 extern void Game_ForceRender(void);
